refactor(rss): extract base64 cursor utility functions
