+++ apps/ethang-rss/src/graphql/util/cursor.test.ts
@@ -1,0 +1,104 @@
+/* eslint-disable unicorn/prefer-uint8array-base64, unicorn/no-this-outside-of-class */
+import { Buffer } from "node:buffer";
+import { describe, expect, it } from "vitest";
+
+import { decodeCursor, encodeCursor } from "./cursor.ts";
+
+describe("cursor utilities", () => {
+  it("should encode and decode a valid cursor", () => {
+    const original = ["2026-06-15T10:00:00.000Z", "item-1"] as [string, string];
+    const encoded = encodeCursor(original);
+    const decoded = decodeCursor(encoded);
+
+    expect(decoded).toEqual(original);
+  });
+
+  it("should encode and decode a cursor with null first value", () => {
+    const original = [null, "item-1"] as [null, string];
+    const encoded = encodeCursor(original);
+    const decoded = decodeCursor(encoded);
+
+    expect(decoded).toEqual(original);
+  });
+
+  it.each([
+    { input: "invalid-base64-string!!" },
+    { input: Buffer.from("invalid-json").toString("base64") },
+    { input: Buffer.from("[]").toString("base64") },
+    { input: Buffer.from("[1]").toString("base64") },
+    { input: Buffer.from("[1, 2, 3]").toString("base64") },
+    { input: Buffer.from('["val1", 2]').toString("base64") },
+    { input: Buffer.from('[2, "val2"]').toString("base64") }
+  ])("should return null for invalid cursor structure: $input", ({ input }) => {
+    const decoded = decodeCursor(input);
+    expect(decoded).toBeNull();
+  });
+
+  describe("native environment paths", () => {
+    it("should cover toBase64 and fromBase64 native paths when they exist", () => {
+      /* eslint-disable no-extend-native */
+      const originalToBase64 = (Uint8Array.prototype as any).toBase64;
+      const originalFromBase64 = (Uint8Array as any).fromBase64;
+
+      Object.defineProperty(Uint8Array.prototype, "toBase64", {
+        configurable: true,
+        value() {
+          return Buffer.from(this).toString("base64");
+        },
+        writable: true
+      });
+
+      Object.defineProperty(Uint8Array, "fromBase64", {
+        configurable: true,
+        value: (base64: string) => {
+          return new Uint8Array(Buffer.from(base64, "base64"));
+        },
+        writable: true
+      });
+
+      try {
+        const original = ["2026-06-15T10:00:00.000Z", "item-1"] as [
+          string,
+          string
+        ];
+        const encoded = encodeCursor(original);
+        const decoded = decodeCursor(encoded);
+        expect(decoded).toEqual(original);
+      } finally {
+        if (undefined === originalToBase64) {
+          delete (Uint8Array.prototype as any).toBase64;
+        } else {
+          (Uint8Array.prototype as any).toBase64 = originalToBase64;
+        }
+
+        if (undefined === originalFromBase64) {
+          delete (Uint8Array as any).fromBase64;
+        } else {
+          (Uint8Array as any).fromBase64 = originalFromBase64;
+        }
+      }
+    });
+
+    it("should cover fromBase64 throwing error on invalid input", () => {
+      const originalFromBase64 = (Uint8Array as any).fromBase64;
+      Object.defineProperty(Uint8Array, "fromBase64", {
+        configurable: true,
+        value: () => {
+          throw new Error("Invalid base64");
+        },
+        writable: true
+      });
+
+      try {
+        const decoded = decodeCursor("invalid!!");
+        expect(decoded).toBeNull();
+      } finally {
+        if (undefined === originalFromBase64) {
+          delete (Uint8Array as any).fromBase64;
+        } else {
+          (Uint8Array as any).fromBase64 = originalFromBase64;
+        }
+      }
+    });
+  });
+});
